Projects/requirements.txt
pandas
kaggle
python-dotenv